[BUGFIX] Sort the dependencies by subfolder, type and name
Without a global sort, a multi-build-tool project displays the dependencies
per build tool, which mixes badly if there are multiple modules: the content
of the root module of the second build tool appears with the content of the
last module of the first build tool. Sorting by type also helps using the
result.